requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
pinecone-client
numpy
transformers
torch